# out/apps/server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "PermiSET API",
    "description": "API for managing users, roles, and permissions"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Retrieve all users",
        "responses": {
          "200": {
            "description": "List of users",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "role": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "email": { "type": "string" },
                "password": { "type": "string" },
                "role": { "type": "string" },
                "status": { "type": "string" }
              },
              "required": ["name", "email", "password", "role", "status"]
            }
          }
        ],
        "responses": {
          "201": { "description": "User successfully created" },
          "400": { "description": "Invalid data" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Retrieve a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "name": { "type": "string" },
                "email": { "type": "string" },
                "role": { "type": "string" },
                "status": { "type": "string" }
              }
            }
          },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update a user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "email": { "type": "string" },
                "role": { "type": "string" },
                "status": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "User successfully updated" },
          "404": { "description": "User not found" }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User ID"
          }
        ],
        "responses": {
          "204": { "description": "User successfully deleted" },
          "404": { "description": "User not found" }
        }
      }
    },
...
  }
}